--- feature/home/src/main/java/com/tgyuu/home/graph/addtodo/ui/bottomsheet/RepeatCycleBottomSheet.kt
@@ -1,89 +1,0 @@
-package com.tgyuu.home.graph.addtodo.ui.bottomsheet
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.heightIn
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material3.Icon
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import com.tgyuu.common.util.clickable
-import com.tgyuu.common.util.verticalScrollbar
-import com.tgyuu.designsystem.component.bottomsheet.EbbingBottomSheetHeader
-import com.tgyuu.designsystem.component.bottomsheet.EbbingBottomSheetListItemDefault
-import com.tgyuu.designsystem.component.EbbingSolidButton
-import com.tgyuu.designsystem.foundation.EbbingTheme
-import com.tgyuu.designsystem.model.RepeatCycleUiModel
-import kotlinx.collections.immutable.ImmutableList
-
-@Composable
-internal fun RepeatCycleBottomSheet(
-    repeatCycleList: ImmutableList<RepeatCycleUiModel>,
-    originRepeatCycle: RepeatCycleUiModel?,
-    updateRepeatCycle: (RepeatCycleUiModel) -> Unit,
-    onAddRepeatCycleClick: () -> Unit,
-) {
-    var newRepeatCycle by remember(originRepeatCycle) { mutableStateOf(originRepeatCycle) }
-    val listState = rememberLazyListState()
-
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 20.dp),
-    ) {
-        EbbingBottomSheetHeader(
-            title = "반복 주기",
-            rightComponent = {
-                Icon(
-                    imageVector = Icons.Default.Add,
-                    contentDescription = null,
-                    tint = EbbingTheme.colors.black,
-                    modifier = Modifier
-                        .size(32.dp)
-                        .clickable { onAddRepeatCycleClick() },
-                )
-            })
-
-        LazyColumn(
-            state = listState,
-            modifier = Modifier
-                .fillMaxWidth()
-                .heightIn(max = 300.dp)
-                .padding(top = 12.dp)
-                .verticalScrollbar(
-                    state = listState,
-                    color = EbbingTheme.colors.light1,
-                ),
-        ) {
-            items(
-                items = repeatCycleList,
-                key = { it.id },
-            ) { cycle ->
-                EbbingBottomSheetListItemDefault(
-                    label = cycle.displayName,
-                    checked = cycle.id == newRepeatCycle?.id,
-                    onChecked = { newRepeatCycle = cycle },
-                )
-            }
-        }
-
-        EbbingSolidButton(
-            label = "적용하기",
-            onClick = { newRepeatCycle?.let { updateRepeatCycle(it) } },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 12.dp, bottom = 10.dp),
-        )
-    }
-}
